+++ packages/utils/src/db/db.spec.ts
@@ -166,7 +166,7 @@ describe(`database tests`, function () {
   })
 
   it('hasPacket', async function () {
-    const packetTag = randomBytes(5)
+    const packetTag = Uint8Array.from(randomBytes(16))
 
     const present = await db.checkAndSetPacketTag(packetTag)