+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 <a name="1.89"></a>
+
 ## [1.89](https://github.com/hoprnet/hoprnet/compare/release/ouagadougou...hoprnet:master)
 
 ---