+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+<a name="1.89"></a>
+
+## [1.89](https://github.com/hoprnet/hoprnet/compare/release/ouagadougou...hoprnet:master)
+
+---
+
 <a name="1.88"></a>
 
-## [1.88](https://github.com/hoprnet/hoprnet/compare/release/lisbon...hoprnet:master) (2022-02-14)
+## [1.88](https://github.com/hoprnet/hoprnet/compare/release/lisbon...hoprnet:release/ouagadougou) (2022-03-18)
 
 ### Changes